==> scripts/2_TCGA_getPatientList.R <==
# Makes list of patient IDs that have all 3 data modalities available (RNAseq, mutation and WSI)
# WSI data corresponds to image embeddings generated by Li et al., 2021 (https://arxiv.org/pdf/2011.08939.pdf)

library(readr)
library(dplyr)
library(stringr)

setwd("/Users/sophie/Dropbox (The Francis Crick)/Sophie/Astra_project/Repositories/TCGA_LUAD_project/")

# Functions ####################################################################
findIntersection <- function(dataframes){
  #' returns the list of patient IDs that are common between the input dataframes 
  #' each dataframe needs to have a column called "cases.submitter_id" with the corresponding patient IDs (e.g."TCGA-44-7660")
  #' @param dataframes: a list of the dataframes to be queried
 
  list <- vector(mode = "list", length = length(dataframes)) # creates empty list to store patient ID lists
  
  for (i in 1:length(dataframes)){
    df <- dataframes[[i]] 
    list[[i]] <- df$cases.submitter_id
  }
  
  patient_list <- Reduce(intersect, list) # finds intersection between patient lists
  patient_list <- data.frame("cases.submitter_id" = patient_list)
  
  return(patient_list)
}

addMutantStatus <- function(patient_list, mut_df, gene){
  #' Adds column with mutation status to patient_list dataframe
  #' @param patient_list a dataframe with list of patient IDs ("cases.submitter_id")
  #' @param mut_df a dataframe with mutation information. Needs to contain a column named "cases.submitter_id" with patient IDs,
  #' as well as "Hugo_Symbol" containing gene symbols
  #' @param gene a string corresponding to the name of the gene of interest (e.g. "STK11")

  # get list of patients that are mutant for the gene of interest
  mut_patients <- mut_df %>% 
    filter(Hugo_Symbol == gene) %>% 
    pull(cases.submitter_id) %>% 
    unique()
  
  # adds mutation status to patient_list
  df <- patient_list %>% 
    mutate(mutation_status = case_when(cases.submitter_id %in% mut_patients ~ 1,
                                       TRUE ~ 0))
  
  return(df)
}

# Load dataframes and add necessary columns ####################################
luad_rna <- read_csv("data/luad_rna.csv")
luad_mut <- read_csv("data/luad_mut_mc3.csv")
luad_wsi <- read_csv("data/tcga-dataset/LUAD.csv")

# adds column with patient ID to dataframe
luad_wsi <- luad_wsi[,1] %>% 
  rename("barcode" = "0") %>% 
  mutate(cases.submitter_id = substr(barcode, 21, 32))

# Find intersection and save output ############################################
dataframes <- list(luad_rna, luad_mut, luad_wsi)
patient_list <- findIntersection(dataframes)
# add mutant status to patient_list dataframe
patient_list <- addMutantStatus(patient_list, luad_mut, "STK11")

write_csv(patient_list, "data/luad_patient_list.csv")
